requirements.txt
annotated-types==0.6.0
anyio==4.2.0
black==23.12.1
click==8.1.7
fastapi==0.108.0
ffmpeg-python==0.2.0
future==0.18.3
h11==0.14.0
httptools==0.6.1
idna==3.6
iniconfig==2.0.0
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.1.0
pluggy==1.3.0
pydantic==2.5.3
pydantic_core==2.14.6
pydub==0.22.0
pytest==7.4.4
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.1
sniffio==1.3.0
starlette==0.32.0.post1
typing_extensions==4.9.0
uvicorn==0.25.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
